refactor: move interstitials into its own file in `src/user/`
